+++ NEWS.md
@@ -1,6 +1,7 @@
 ## v5.0.1 2024.03.25
 ### Bug fixes
 * Browse Config: fix exception while trying to validate empty type ids ([MSEARCH-998](https://folio-org.atlassian.net/browse/MSEARCH-998))
+* ECS: Modify filters for sub-resources during browse ([MSEARCH-1000](https://folio-org.atlassian.net/browse/MSEARCH-1000))
 
 ---